pyproject.toml
[project]
name = "chat-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "cryptography>=45.0.5",
    "fastapi[standard]>=0.116.1",
    "google-search-results>=2.4.2",
    "openai>=1.98.0",
    "pymysql>=1.1.1",
    "python-dotenv>=1.1.1",
    "sqlmodel>=0.0.24",
]
